test(mlx): consolidate tokenizer compatibility coverage
